refactor: new middleware for exceptions errors
